483148: Move buildLink helper to view folder
