+++ src/server/admin/permissions/transformers/scope-teams-to-task-list.js
@@ -1,4 +1,4 @@
-import { buildLink } from '~/src/server/common/helpers/build-link.js'
+import { buildLink } from '~/src/server/common/helpers/view/build-link.js'
 
 function transformScopeTeamsToTaskList(scope, withActions = true) {
   const items = scope.teams.map((team) => {
