+++ src/server/admin/permissions/transformers/scope-users-to-task-list.js
@@ -1,4 +1,4 @@
-import { buildLink } from '~/src/server/common/helpers/build-link.js'
+import { buildLink } from '~/src/server/common/helpers/view/build-link.js'
 
 function transformScopeUsersToTaskList(scope, withActions = true) {
   const items = scope.users.map((user) => {
