+++ src/server/admin/teams/transformers/team-users-to-task-list.js
@@ -1,4 +1,4 @@
-import { buildLink } from '~/src/server/common/helpers/build-link.js'
+import { buildLink } from '~/src/server/common/helpers/view/build-link.js'
 
 function transformTeamUsersToTaskList(team, withActions = true) {
   const items = team.users.map((user) => {
