+++ src/server/admin/users/transformers/transform-summary-user-rows.js
@@ -1,7 +1,7 @@
 import isNull from 'lodash/isNull.js'
 
 import { noValue } from '~/src/server/common/constants/no-value.js'
-import { buildLink } from '~/src/server/common/helpers/build-link.js'
+import { buildLink } from '~/src/server/common/helpers/view/build-link.js'
 
 function buildRow(name, value, stepPath, query, queryValue = value) {
   const queryString =
