+++ src/server/admin/users/transformers/user-scopes-to-task-list.js
@@ -1,4 +1,4 @@
-import { buildLink } from '~/src/server/common/helpers/build-link.js'
+import { buildLink } from '~/src/server/common/helpers/view/build-link.js'
 
 function transformUserScopesToTaskList(team) {
   const items =
