+++ src/server/common/helpers/view/build-link.js
