+++ src/server/create/repository/controllers/success.js
@@ -1,6 +1,6 @@
 import { config } from '~/src/config/config.js'
 import { provideCreate } from '~/src/server/create/helpers/pre/provide-create.js'
-import { buildLink } from '~/src/server/common/helpers/build-link.js'
+import { buildLink } from '~/src/server/common/helpers/view/build-link.js'
 import { noSessionRedirect } from '~/src/server/create/helpers/ext/no-session-redirect.js'
 import { setStepComplete } from '~/src/server/create/helpers/form/index.js'
