+++ src/server/services/about/transformers/service-status.js
@@ -3,7 +3,7 @@ import { unknownValue } from '~/src/server/common/constants/no-value.js'
 import { statusTagClassMap } from '~/src/server/common/helpers/status-tag-class-map.js'
 import { removeUrlParts } from '~/src/server/common/helpers/remove-url-parts.js'
 import { creationStatuses } from '~/src/server/common/constants/creation-statuses.js'
-import { buildLink } from '~/src/server/common/helpers/build-link.js'
+import { buildLink } from '~/src/server/common/helpers/view/build-link.js'
 
 const githubOrg = config.get('githubOrg')
