+++ src/server/teams/transformers/to-task-list.js
@@ -1,4 +1,4 @@
-import { buildLink } from '~/src/server/common/helpers/build-link.js'
+import { buildLink } from '~/src/server/common/helpers/view/build-link.js'
 import { config } from '~/src/config/config.js'
 
 function transformToTaskList(repositories = []) {
